+++ src/Print3dServer.Core/Usings.cs
@@ -1,4 +1,3 @@
 ﻿global using CommunityToolkit.Mvvm.ComponentModel; 
 global using RestSharp;
-//global using Newtonsoft.Json;
 global using System.Xml.Serialization;
